+++ users/rhruiz/rgblight/rgblight.h
@@ -2,10 +2,10 @@
 #ifdef RGBLIGHT_ENABLE
 /* underglow control */
 void rhruiz_rgblight_reset(void);
-void rhruiz_change_leds_to(uint16_t, uint8_t);
+void rhruiz_change_leds_to(uint8_t, uint8_t);
 #    ifndef RGBLIGHT_LAYERS
 typedef struct {
-    uint16_t hue;
+    uint8_t hue;
     uint8_t  sat;
 } hue_sat_pair;
 #    endif
